+++ solution/backend/regulations/sitemap.py
@@ -3,7 +3,7 @@ from datetime import datetime
 from django.urls import reverse
 
 from regcore.models import Part
-from resources.models import AbstractResource
+from resources.models import Resource
 
 
 class PartSitemap(Sitemap):
@@ -63,7 +63,7 @@ class SupplementalContentSitemap(Sitemap):
     priority = 0.5
 
     def items(self):
-        return AbstractResource.objects.filter(approved=True)
+        return Resource.objects.filter(approved=True)
 
     def lastmod(self, item):
         return item.updated_at
