+++ solution/backend/regulations/views/supplemental_content.py
@@ -2,7 +2,7 @@ from django.views.generic.base import TemplateView
 from django.http import Http404
 from django.urls import reverse
 
-from resources.models import AbstractResource
+from resources.models import Resource
 
 
 class SupplementalContentView(TemplateView):
@@ -13,7 +13,7 @@ class SupplementalContentView(TemplateView):
         context = super().get_context_data(**kwargs)
         content_id = self.kwargs.get("id")
         try:
-            content = AbstractResource.objects.all().select_subclasses().get(id=content_id, approved=True)
+            content = Resource.objects.all().get(id=content_id, approved=True)
             c = {
                 "mod_time": content.updated_at.isoformat(),
                 "title": getattr(content, "name", "n/a"),
@@ -22,5 +22,5 @@ class SupplementalContentView(TemplateView):
                 "redirect_link": getattr(content, "url", reverse("homepage")),
             }
             return {**context, **c}
-        except AbstractResource.DoesNotExist:
+        except Resource.DoesNotExist:
             raise Http404
